added multiple patterns capability to parser
added parser level checking of weather any match is in progress
new function _update_match_state

adjusted Patter test_match function to save the value to its
_match_state variable, used by Parser._update_match_state

refactored to return the match value for each pattern per character
so that it can be added or not in _run_char depending on if
a match is in progress.

fix for null end call, removing null call, character count for match
now accurate length count was off

added a _head variable to parser to keep track of last character
in the stream appended to the content. used mostly for debugging.