Refactored classes to use prototype chain and instanceof error checking
moved all the loping over properties in to function copyprops

NAMEInterface classes are used to create new objects, don't do
anything in their constructors, so they can be used in other
objects prototype chains.

    var CLASSInterface = function(){}
    CLASSInterface.prototype = {
        func: function(){
            return 'func function called'
        }

    var CLASS = function(){
            //... constructor code for an instance
        };
        Class.prototype = new CLASSInterface;

this allows typechecking with instanceof CLASSInterface
because CLASSInterface will be in the prototype chain.

    CLASS instanceof CLASSInterface // true

the properties of CLASSInterface will not be in CLASS
but they will be accesible through the prototype chain

    cls = new CLASS()
    for(prop in cls)
        print(prop); // nothing will print

but it is accessible on the object because of the
prototype chain

    cls = new CLASS();
    cls.func // returns 'func function called'