rework of filter_by_state, in precedual, fix for dumping matches in place
Created a pending_stack in state manager that pushes, pops, and shifts
matches on and off, making the longest matching pattern accessible
as it will always be at the top

created increadible debugging function for viewing the patterns
beeing tracked in the pending_heap, available in Parser.parse_debug

removing some prototype level properties as they accidently behave
like static class level variables, now created in objects constructor
functions.

result object now contains the portion of the stream it contains,
this is neccessary because the content is an altered version
of the stream characters it represents.

// result object
[statuscode, content, portionlength]
renamed PatternInterface from get to handle