Created new wordboundary based pattern TagWordPattern, and adjustments
conclude(): new function in PatternInterface, RegionInterface, and PendingStack
called in when end of document is reached, runs similar functionality
to evaluate_pattern, moves pending patterns into content stack and
then processes those values.

This was necessary for the exception when a word boundary is the end of
the document.

word patterns in the javascript_syntax.js file move to TagWordPattern.

All patterns now have an id attribute _id, turned out not to be
necessary, but it's a nice addition to decentralize the pattern_index,
so it's included in this commit.